Let PlaceholderModifier use Retokenizer implementation for now
Should be mostly compatible, except that the end-to-end tests are failing? Did I change the order of random calls?
